working on fitting modules
